CoDICE: Code Clean up and pre-step refactor
