+++ imap_processing/codice/constants.py
@@ -19,712 +19,13 @@ import numpy as np
 
 from imap_processing.codice.utils import CODICEAPID, CoDICECompression
 
-# Grouping of APIDs used to signify similar L1a processing
-APIDS_FOR_SCIENCE_PROCESSING = [
-    CODICEAPID.COD_HI_INST_COUNTS_AGGREGATED,
-    CODICEAPID.COD_HI_INST_COUNTS_PRIORITIES,
-    CODICEAPID.COD_HI_INST_COUNTS_SINGLES,
-    CODICEAPID.COD_HI_SECT_SPECIES_COUNTS,
-    CODICEAPID.COD_LO_INST_COUNTS_AGGREGATED,
-    CODICEAPID.COD_LO_INST_COUNTS_SINGLES,
-    CODICEAPID.COD_LO_SW_ANGULAR_COUNTS,
-    CODICEAPID.COD_LO_NSW_ANGULAR_COUNTS,
-    CODICEAPID.COD_LO_SW_PRIORITY_COUNTS,
-    CODICEAPID.COD_LO_NSW_PRIORITY_COUNTS,
-    CODICEAPID.COD_LO_SW_SPECIES_COUNTS,
-    CODICEAPID.COD_LO_NSW_SPECIES_COUNTS,
-]
-
-# Mapping between descriptors and APID
-CODICEAPID_MAPPING = {
-    "hskp": CODICEAPID.COD_NHK,
-    "lo-ialirt": CODICEAPID.COD_LO_IAL,
-    "lo-direct-events": CODICEAPID.COD_LO_PHA,
-    "lo-sw-priority": CODICEAPID.COD_LO_SW_PRIORITY_COUNTS,
-    "lo-sw-species": CODICEAPID.COD_LO_SW_SPECIES_COUNTS,
-    "lo-nsw-species": CODICEAPID.COD_LO_NSW_SPECIES_COUNTS,
-    "lo-sw-angular": CODICEAPID.COD_LO_SW_ANGULAR_COUNTS,
-    "lo-nsw-angular": CODICEAPID.COD_LO_NSW_ANGULAR_COUNTS,
-    "lo-nsw-priority": CODICEAPID.COD_LO_NSW_PRIORITY_COUNTS,
-    "lo-counters-aggregated": CODICEAPID.COD_LO_INST_COUNTS_AGGREGATED,
-    "lo-counters-singles": CODICEAPID.COD_LO_INST_COUNTS_SINGLES,
-    "hi-ialirt": CODICEAPID.COD_HI_IAL,
-    "hi-direct-events": CODICEAPID.COD_HI_PHA,
-    "hi-counters-aggregated": CODICEAPID.COD_HI_INST_COUNTS_AGGREGATED,
-    "hi-counters-singles": CODICEAPID.COD_HI_INST_COUNTS_SINGLES,
-    "hi-omni": CODICEAPID.COD_HI_OMNI_SPECIES_COUNTS,
-    "hi-sectored": CODICEAPID.COD_HI_SECT_SPECIES_COUNTS,
-    "hi-priority": CODICEAPID.COD_HI_INST_COUNTS_PRIORITIES,
-}
-
+# -------L1A Constants-------
 # Numerical constants
 SPIN_PERIOD_CONVERSION = 0.00032
 K_FACTOR = 5.76  # This is used to convert voltages to energies in L2
-HI_ACQUISITION_TIME = 0.59916
 NUM_ESA_STEPS = 128
 LO_DESPIN_SPIN_SECTORS = 24
-
-# CDF variable names used for lo data products
-LO_COUNTERS_SINGLES_VARIABLE_NAMES = ["apd_singles"]
-LO_SW_ANGULAR_VARIABLE_NAMES = ["hplus", "heplusplus", "oplus6", "fe_loq"]
-LO_NSW_ANGULAR_VARIABLE_NAMES = ["heplusplus"]
-LO_SW_PRIORITY_VARIABLE_NAMES = [
-    "p0_tcrs",
-    "p1_hplus",
-    "p2_heplusplus",
-    "p3_heavies",
-    "p4_dcrs",
-]
-LO_NSW_PRIORITY_VARIABLE_NAMES = ["p5_heavies", "p6_hplus_heplusplus"]
-LO_SW_SPECIES_VARIABLE_NAMES = [
-    "hplus",
-    "heplusplus",
-    "cplus4",
-    "cplus5",
-    "cplus6",
-    "oplus5",
-    "oplus6",
-    "oplus7",
-    "oplus8",
-    "ne",
-    "mg",
-    "si",
-    "fe_loq",
-    "fe_hiq",
-    "heplus",
-    "cnoplus",
-]
-LO_SW_SOLAR_WIND_SPECIES_VARIABLE_NAMES = [
-    "hplus",
-    "heplusplus",
-    "cplus4",
-    "cplus5",
-    "cplus6",
-    "oplus5",
-    "oplus6",
-    "oplus7",
-    "oplus8",
-    "ne",
-    "mg",
-    "si",
-    "fe_loq",
-    "fe_hiq",
-]
-LO_SW_PICKUP_ION_SPECIES_VARIABLE_NAMES = [
-    "heplus",
-    "cnoplus",
-]
-LO_NSW_SPECIES_VARIABLE_NAMES = [
-    "hplus",
-    "heplusplus",
-    "c",
-    "o",
-    "ne_si_mg",
-    "fe",
-    "heplus",
-    "cnoplus",
-]
-LO_IALIRT_VARIABLE_NAMES = [
-    "heplusplus",
-    "cplus5",
-    "cplus6",
-    "oplus6",
-    "oplus7",
-    "oplus8",
-    "mg",
-    "fe_loq",
-    "fe_hiq",
-]
-
-# CDF variable names used for hi data products
-HI_COUNTERS_SINGLES_VARIABLE_NAMES = ["tcr", "ssdo", "stssd"]
-HI_OMNI_VARIABLE_NAMES = ["h", "he3", "he4", "c", "o", "ne_mg_si", "fe", "uh", "junk"]
-HI_PRIORITY_VARIABLE_NAMES = [
-    "Priority0",
-    "Priority1",
-    "Priority2",
-    "Priority3",
-    "Priority4",
-    "Priority5",
-]
-HI_SECTORED_VARIABLE_NAMES = ["h", "he3he4", "cno", "fe"]
-HI_IALIRT_VARIABLE_NAMES = ["h"]
-
-# Final I-ALiRT data product fields
-CODICE_LO_IAL_DATA_FIELDS = [
-    "c_over_o_abundance",
-    "mg_over_o_abundance",
-    "fe_over_o_abundance",
-    "c_plus_6_over_c_plus_5_ratio",
-    "o_plus_7_over_o_plus_6_ratio",
-    "fe_low_over_fe_high_ratio",
-]
-CODICE_HI_IAL_DATA_FIELDS = ["h"]
-
-# lo- and hi-counters-aggregated data product variables are dynamically
-# determined based on the number of active counters
-LO_COUNTERS_AGGREGATED_ACTIVE_VARIABLES = {
-    "tcr": True,
-    "dcr": True,
-    "tof_plus_apd": False,
-    "tof_only": False,
-    "position_plus_apd": False,
-    "position_only": False,
-    "sta_or_stb_plus_apd": False,
-    "sta_or_stb_only": False,
-    "reserved1": False,
-    "reserved2": False,
-    "sp_only": False,
-    "apd_only": False,
-    "low_tof_cutoff": False,
-    "sta": True,
-    "stb": True,
-    "sp": True,
-    "total_position_count": True,
-    "invalid_position_count": False,
-    "asic1_flag_invalid": False,
-    "asic2_flag_invalid": False,
-    "asic1_channel_invalid": False,
-    "asic2_channel_invalid": False,
-    "tec4_timeout_tof_no_pos": False,
-    "tec4_timeout_pos_no_tof": False,
-    "tec4_timeout_no_pos_tof": False,
-    "tec5_timeout_tof_no_pos": False,
-    "tec5_timeout_pos_no_tof": False,
-    "tec5_timeout_no_pos_tof": False,
-}
-LO_COUNTERS_AGGREGATED_VARIABLE_NAMES = [
-    name
-    for name, is_active in LO_COUNTERS_AGGREGATED_ACTIVE_VARIABLES.items()
-    if is_active
-]
-HI_COUNTERS_AGGREGATED_ACTIVE_VARIABLES = {
-    "dcr": True,
-    "sto": True,
-    "spo": True,
-    "reserved1": False,
-    "mst": True,
-    "ssdo": True,
-    "stssd": True,
-    "reserved4": False,
-    "reserved5": False,
-    "low_tof_cutoff": True,
-    "reserved6": False,
-    "reserved7": False,
-    "asic1_flag_invalid": True,
-    "asic2_flag_invalid": True,
-    "asic1_channel_invalid": True,
-    "asic2_channel_invalid": True,
-}
-HI_COUNTERS_AGGREGATED_VARIABLE_NAMES = [
-    name
-    for name, is_active in HI_COUNTERS_AGGREGATED_ACTIVE_VARIABLES.items()
-    if is_active
-]
-
-# List of data products that require application of despinning algorithm
-REQUIRES_DESPINNING = [
-    "imap_codice_l1a_lo-sw-angular",
-    "imap_codice_l1a_lo-nsw-angular",
-    # TBD if this requires despinning
-    # "imap_codice_l1a_lo-sw-priority",
-    # "imap_codice_l1a_lo-nsw-priority",
-]
-
-# Energy tables for CoDICE-Hi data products. These values represent the edges
-# of the bins, and are used in the CoDICE L1a pipeline to compute the centers
-# and deltas of the bins, which then get stored in the CDF files for future use.
-# These are defined in the "Data Products - Hi" tab of the "*-SCI-LUT-*.xml"
-# spreadsheet that largely defines CoDICE processing.
-IALIRT_ENERGY_TABLE = {
-    "h": [
-        0.05,
-        0.070710678,
-        0.1,
-        0.141421356,
-        0.2,
-        0.282842712,
-        0.4,
-        0.565685425,
-        0.8,
-        1.13137085,
-        1.6,
-        2.2627417,
-        3.2,
-        4.5254834,
-        6.4,
-        9.050966799,
-    ],
-}
-
-OMNI_ENERGY_TABLE = {
-    "h": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-    ],
-    "he3": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-    ],
-    "he4": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-    ],
-    "c": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-        5.1200000000,
-        7.2407734394,
-        10.2400000000,
-    ],
-    "o": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-        5.1200000000,
-        7.2407734394,
-        10.2400000000,
-    ],
-    "ne_mg_si": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-    ],
-    "fe": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-        0.1600000000,
-        0.2262741700,
-        0.3200000000,
-        0.4525483400,
-        0.6400000000,
-        0.9050966799,
-        1.2800000000,
-        1.8101933598,
-        2.5600000000,
-        3.6203867197,
-        5.1200000000,
-        7.2407734394,
-        10.2400000000,
-    ],
-    "uh": [
-        0.0200000000,
-        0.0282842712,
-        0.0400000000,
-        0.0565685425,
-        0.0800000000,
-        0.1131370850,
-    ],
-    "junk": [0.0200000000, 0.0282842712],
-}
-
-# In the future, we get csv file with these column:
-# species, min_energy, max_energy, product (descriptor of the product)
-SECTORED_ENERGY_TABLE = {
-    "h": [0.020, 0.040, 0.080, 0.160, 0.320, 0.640, 1.280, 2.560, 5.120],
-    "he3he4": [0.020, 0.040, 0.080, 0.160, 0.320, 0.640, 1.280, 2.560, 5.120],
-    "cno": [0.020, 0.040, 0.080, 0.160, 0.320, 0.640, 1.280, 2.560, 5.120],
-    "fe": [0.020, 0.040, 0.080, 0.160, 0.320, 0.640, 1.280, 2.560, 5.120],
-}
-
-# Various configurations to support processing of individual data products
-# Much of these are described in the algorithm document in chapter 10 ("Data
-# Level 1A")
-DATA_PRODUCT_CONFIGURATIONS: dict[CODICEAPID | int, dict] = {
-    CODICEAPID.COD_HI_IAL: {
-        "dataset_name": "imap_codice_l1a_hi-ialirt",
-        "energy_table": IALIRT_ENERGY_TABLE,
-        "dims": {"esa_step": 15, "inst_az": 4},
-        "instrument": "hi",
-        "num_counters": 1,
-        "num_spins": 4,
-        "support_variables": [
-            "data_quality",
-            "spin_period",
-            "energy_h",
-        ],
-        "variable_names": HI_IALIRT_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_HI_INST_COUNTS_AGGREGATED: {
-        "dataset_name": "imap_codice_l1a_hi-counters-aggregated",
-        "dims": {},
-        "instrument": "hi",
-        "num_counters": len(
-            HI_COUNTERS_AGGREGATED_VARIABLE_NAMES
-        ),  # The number of counters depends on the number of *active* counters
-        "support_variables": ["data_quality", "spin_period"],
-        "variable_names": HI_COUNTERS_AGGREGATED_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_HI_INST_COUNTS_SINGLES: {
-        "dataset_name": "imap_codice_l1a_hi-counters-singles",
-        "dims": {
-            "ssd_index": 12,
-        },
-        "instrument": "hi",
-        "num_counters": 3,
-        "support_variables": ["data_quality", "spin_period"],
-        "variable_names": HI_COUNTERS_SINGLES_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_HI_INST_COUNTS_PRIORITIES: {
-        "dataset_name": "imap_codice_l1a_hi-priority",
-        "dims": {},
-        "instrument": "hi",
-        "num_counters": 6,
-        "support_variables": ["data_quality", "spin_period"],
-        "variable_names": HI_PRIORITY_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_HI_OMNI_SPECIES_COUNTS: {
-        "dataset_name": "imap_codice_l1a_hi-omni",
-        "energy_table": OMNI_ENERGY_TABLE,
-        "dims": {"esa_step": 15, "inst_az": 4},
-        "instrument": "hi",
-        "num_counters": 8,
-        "num_spins": 4,
-        "support_variables": [
-            "data_quality",
-            "spin_period",
-            "energy_h",
-            "energy_he3",
-            "energy_he4",
-            "energy_c",
-            "energy_o",
-            "energy_ne_mg_si",
-            "energy_fe",
-            "energy_uh",
-            "energy_junk",
-        ],
-        "variable_names": HI_OMNI_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_HI_SECT_SPECIES_COUNTS: {
-        "dataset_name": "imap_codice_l1a_hi-sectored",
-        "energy_table": SECTORED_ENERGY_TABLE,
-        "dims": {
-            "esa_step": 8,
-            "ssd_index": 12,
-            "spin_sector_index": 12,
-        },
-        "instrument": "hi",
-        "num_counters": 4,
-        "num_spins": 16,
-        "support_variables": [
-            "data_quality",
-            "spin_period",
-            "energy_h",
-            "energy_he3he4",
-            "energy_cno",
-            "energy_fe",
-        ],
-        "variable_names": HI_SECTORED_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_IAL: {
-        "dataset_name": "imap_codice_l1a_lo-ialirt",
-        "dims": {"esa_step": 128, "spin_sector": 1},
-        "instrument": "lo",
-        "num_counters": 9,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_IALIRT_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_INST_COUNTS_AGGREGATED: {
-        "dataset_name": "imap_codice_l1a_lo-counters-aggregated",
-        "dims": {"esa_step": 128, "spin_sector_pairs": 6},
-        "instrument": "lo",
-        "num_counters": len(
-            LO_COUNTERS_AGGREGATED_VARIABLE_NAMES
-        ),  # The number of counters depends on the number of *active* counters
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_COUNTERS_AGGREGATED_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_INST_COUNTS_SINGLES: {
-        "dataset_name": "imap_codice_l1a_lo-counters-singles",
-        "dims": {"esa_step": 128, "inst_az": 24, "spin_sector_pairs": 6},
-        "instrument": "lo",
-        "num_counters": 1,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_COUNTERS_SINGLES_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_SW_ANGULAR_COUNTS: {
-        "dataset_name": "imap_codice_l1a_lo-sw-angular",
-        "dims": {"esa_step": 128, "inst_az": 5, "spin_sector": 12},
-        "instrument": "lo",
-        "num_counters": 4,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_SW_ANGULAR_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_NSW_ANGULAR_COUNTS: {
-        "dataset_name": "imap_codice_l1a_lo-nsw-angular",
-        "dims": {"esa_step": 128, "inst_az": 19, "spin_sector": 12},
-        "instrument": "lo",
-        "num_counters": 1,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_NSW_ANGULAR_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_SW_PRIORITY_COUNTS: {
-        "dataset_name": "imap_codice_l1a_lo-sw-priority",
-        "dims": {"esa_step": 128, "spin_sector": 12},
-        "instrument": "lo",
-        "num_counters": 5,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_SW_PRIORITY_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_NSW_PRIORITY_COUNTS: {
-        "dataset_name": "imap_codice_l1a_lo-nsw-priority",
-        "dims": {"esa_step": 128, "spin_sector": 12},
-        "instrument": "lo",
-        "num_counters": 2,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_NSW_PRIORITY_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_SW_SPECIES_COUNTS: {
-        "dataset_name": "imap_codice_l1a_lo-sw-species",
-        "dims": {"esa_step": 128, "spin_sector": 1},
-        "instrument": "lo",
-        "num_counters": 16,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_SW_SPECIES_VARIABLE_NAMES,
-    },
-    CODICEAPID.COD_LO_NSW_SPECIES_COUNTS: {
-        "dataset_name": "imap_codice_l1a_lo-nsw-species",
-        "dims": {"esa_step": 128, "spin_sector": 1},
-        "instrument": "lo",
-        "num_counters": 8,
-        "support_variables": [
-            "energy_table",
-            "acquisition_time_per_step",
-            "rgfo_half_spin",
-            "nso_half_spin",
-            "sw_bias_gain_mode",
-            "st_bias_gain_mode",
-            "data_quality",
-            "spin_period",
-            "k_factor",
-        ],
-        "variable_names": LO_NSW_SPECIES_VARIABLE_NAMES,
-    },
-}
-
-# Various configurations to support L1b processing of individual data products
-# Much of these are described in the algorithm document in chapter 11 ("Data
-# Level 1B")
-L1B_DATA_PRODUCT_CONFIGURATIONS: dict[str, dict] = {
-    "hi-counters-aggregated": {
-        "num_spin_sectors": 24,
-        "num_spins": 16,
-    },
-    "hi-counters-singles": {
-        "num_spin_sectors": 24,
-        "num_spins": 16,
-    },
-    "hi-ialirt": {
-        "num_spin_sectors": 6,
-        "num_spins": 4,
-    },
-    "hi-omni": {
-        "num_spin_sectors": 24,
-        "num_spins": 4,
-    },
-    "hi-priority": {
-        "num_spin_sectors": 24,
-        "num_spins": 16,
-    },
-    "hi-sectored": {
-        "num_spin_sectors": 2,
-        "num_spins": 16,
-    },
-    "lo-counters-aggregated": {
-        "num_spin_sectors": 2,
-    },
-    "lo-counters-singles": {
-        "num_spin_sectors": 2,
-    },
-    "lo-nsw-angular": {
-        "num_spin_sectors": 1,
-    },
-    "lo-sw-angular": {
-        "num_spin_sectors": 1,
-    },
-    "lo-nsw-priority": {
-        "num_spin_sectors": 1,
-    },
-    "lo-sw-priority": {
-        "num_spin_sectors": 1,
-    },
-    "lo-nsw-species": {
-        "num_spin_sectors": 12,
-    },
-    "lo-sw-species": {
-        "num_spin_sectors": 12,
-    },
-    "lo-ialirt": {
-        "num_spin_sectors": 12,
-    },
-}
-
+MAX_DE_EVENTS_PER_PACKET = 10000
 
 # Define the packet fields needed to be stored in segmented data and their
 # corresponding bit lengths for direct event data products
@@ -742,8 +43,6 @@ DE_METADATA_FIELDS = {
     "num_events": 32,
     "byte_count": 32,
 }
-
-MAX_DE_EVENTS_PER_PACKET = 10000
 
 # Various configurations to support processing of direct events data products
 # These are described in the algorithm document in chapter 10 ("Data Level 1A")
@@ -865,30 +164,6 @@ DE_DATA_PRODUCT_CONFIGURATIONS: dict[Any, dict[str, Any]] = {
     },
 }
 
-# Define the packet fields needed to be stored in segmented data and their
-# corresponding bit lengths for I-ALiRT data products
-IAL_BIT_STRUCTURE = {
-    "SHCOARSE": 32,
-    "PACKET_VERSION": 16,
-    "SPIN_PERIOD": 16,
-    "ACQ_START_SECONDS": 32,
-    "ACQ_START_SUBSECONDS": 20,
-    "SPARE_00": 8,
-    "ST_BIAS_GAIN_MODE": 2,
-    "SW_BIAS_GAIN_MODE": 2,
-    "TABLE_ID": 32,
-    "PLAN_ID": 16,
-    "PLAN_STEP": 4,
-    "VIEW_ID": 4,
-    "RGFO_HALF_SPIN": 6,
-    "NSO_HALF_SPIN": 6,
-    "SPARE_01": 1,
-    "SUSPECT": 1,
-    "COMPRESSION": 3,
-    "BYTE_COUNT": 23,
-}
-
-
 # Compression ID lookup tables
 # The key is the view_id and the value is the ID for the compression algorithm
 # (see utils.CoDICECompression to see how the values correspond)
@@ -916,91 +191,6 @@ HI_COMPRESSION_ID_LOOKUP = {
     7: CoDICECompression.LOSSY_A_LOSSLESS,
     8: CoDICECompression.LOSSY_A_LOSSLESS,
     9: CoDICECompression.LOSSY_A_LOSSLESS,
-}
-
-# ESA Sweep table ID lookup table
-# The combination of plan_id and plan_step determine the ESA sweep Table to use
-# Currently, ESA sweep table 0 is used for every plan_id/plan_step combination,
-# but may change in the future. These are defined in the "ESA Sweep" tab of the
-# "*-SCI-LUT-*.xml" spreadsheet that largely defines CoDICE processing.
-ESA_SWEEP_TABLE_ID_LOOKUP = {
-    (0, 0): 0,
-    (0, 1): 0,
-    (0, 2): 0,
-    (0, 3): 0,
-    (1, 0): 0,
-    (1, 1): 0,
-    (1, 2): 0,
-    (1, 3): 0,
-    (2, 0): 0,
-    (2, 1): 0,
-    (2, 2): 0,
-    (2, 3): 0,
-    (3, 0): 0,
-    (3, 1): 0,
-    (3, 2): 0,
-    (3, 3): 0,
-    (4, 0): 0,
-    (4, 1): 0,
-    (4, 2): 0,
-    (4, 3): 0,
-    (5, 0): 0,
-    (5, 1): 0,
-    (5, 2): 0,
-    (5, 3): 0,
-    (6, 0): 0,
-    (6, 1): 0,
-    (6, 2): 0,
-    (6, 3): 0,
-    (7, 0): 0,
-    (7, 1): 0,
-    (7, 2): 0,
-    (7, 3): 0,
-}
-
-# Lo Stepping table ID lookup table
-# The combination of plan_id and plan_step determine the Lo Stepping Table to
-# use. Currently, LO Stepping table 0 is used for every plan_id/plan_step
-# combination, but may change in the future. These are defined in the "Lo
-# Stepping" tab of the "*-SCI-LUT-*.xml" spreadsheet that largely defines CoDICE
-# processing. Eg.
-#   (plan_id, plan_step) -> id of acquisition time
-#   (0, 0) -> 0
-
-
-LO_STEPPING_TABLE_ID_LOOKUP = {
-    (0, 0): 0,
-    (0, 1): 0,
-    (0, 2): 0,
-    (0, 3): 0,
-    (1, 0): 0,
-    (1, 1): 0,
-    (1, 2): 0,
-    (1, 3): 0,
-    (2, 0): 0,
-    (2, 1): 0,
-    (2, 2): 0,
-    (2, 3): 0,
-    (3, 0): 0,
-    (3, 1): 0,
-    (3, 2): 0,
-    (3, 3): 0,
-    (4, 0): 0,
-    (4, 1): 0,
-    (4, 2): 0,
-    (4, 3): 0,
-    (5, 0): 0,
-    (5, 1): 0,
-    (5, 2): 0,
-    (5, 3): 0,
-    (6, 0): 0,
-    (6, 1): 0,
-    (6, 2): 0,
-    (6, 3): 0,
-    (7, 0): 0,
-    (7, 1): 0,
-    (7, 2): 0,
-    (7, 3): 0,
 }
 
 # Lookup tables for Lossy decompression algorithms "A" and "B"
@@ -1525,6 +715,122 @@ LOSSY_B_TABLE = {
     255: 2151415807,
 }
 
+# ------L1B Constants------
+HI_ACQUISITION_TIME = 0.59916
+
+# TODO: in the future, read from sci-lut
+LO_SW_ANGULAR_VARIABLE_NAMES = ["hplus", "heplusplus", "oplus6", "fe_loq"]
+LO_NSW_ANGULAR_VARIABLE_NAMES = ["heplusplus"]
+LO_NSW_PRIORITY_VARIABLE_NAMES = ["p5_heavies", "p6_hplus_heplusplus"]
+LO_SW_SPECIES_VARIABLE_NAMES = [
+    "hplus",
+    "heplusplus",
+    "cplus4",
+    "cplus5",
+    "cplus6",
+    "oplus5",
+    "oplus6",
+    "oplus7",
+    "oplus8",
+    "ne",
+    "mg",
+    "si",
+    "fe_loq",
+    "fe_hiq",
+    "heplus",
+    "cnoplus",
+]
+
+# Various configurations to support L1b processing of individual data products
+# Much of these are described in the algorithm document in chapter 11 ("Data
+# Level 1B")
+L1B_DATA_PRODUCT_CONFIGURATIONS: dict[str, dict] = {
+    "hi-counters-aggregated": {
+        "num_spin_sectors": 24,
+        "num_spins": 16,
+    },
+    "hi-counters-singles": {
+        "num_spin_sectors": 24,
+        "num_spins": 16,
+    },
+    "hi-ialirt": {
+        "num_spin_sectors": 6,
+        "num_spins": 4,
+    },
+    "hi-omni": {
+        "num_spin_sectors": 24,
+        "num_spins": 4,
+    },
+    "hi-priority": {
+        "num_spin_sectors": 24,
+        "num_spins": 16,
+    },
+    "hi-sectored": {
+        "num_spin_sectors": 2,
+        "num_spins": 16,
+    },
+    "lo-counters-aggregated": {
+        "num_spin_sectors": 2,
+    },
+    "lo-counters-singles": {
+        "num_spin_sectors": 2,
+    },
+    "lo-nsw-angular": {
+        "num_spin_sectors": 1,
+    },
+    "lo-sw-angular": {
+        "num_spin_sectors": 1,
+    },
+    "lo-nsw-priority": {
+        "num_spin_sectors": 1,
+    },
+    "lo-sw-priority": {
+        "num_spin_sectors": 1,
+    },
+    "lo-nsw-species": {
+        "num_spin_sectors": 12,
+    },
+    "lo-sw-species": {
+        "num_spin_sectors": 12,
+    },
+    "lo-ialirt": {
+        "num_spin_sectors": 12,
+    },
+}
+
+# ------L2 Constants------
+LO_SW_SOLAR_WIND_SPECIES_VARIABLE_NAMES = [
+    "hplus",
+    "heplusplus",
+    "cplus4",
+    "cplus5",
+    "cplus6",
+    "oplus5",
+    "oplus6",
+    "oplus7",
+    "oplus8",
+    "ne",
+    "mg",
+    "si",
+    "fe_loq",
+    "fe_hiq",
+]
+LO_SW_PICKUP_ION_SPECIES_VARIABLE_NAMES = [
+    "heplus",
+    "cnoplus",
+]
+LO_NSW_SPECIES_VARIABLE_NAMES = [
+    "hplus",
+    "heplusplus",
+    "c",
+    "o",
+    "ne_si_mg",
+    "fe",
+    "heplus",
+    "cnoplus",
+]
+HI_OMNI_VARIABLE_NAMES = ["h", "he3", "he4", "c", "o", "ne_mg_si", "fe", "uh", "junk"]
+HI_SECTORED_VARIABLE_NAMES = ["h", "he3he4", "cno", "fe"]
 # Lookup table for CoDICE-Lo despinning pixel orientations
 # See section 9.3.4 of the algorithm document for further information
 PIXEL_ORIENTATIONS = {
@@ -1657,563 +963,6 @@ PIXEL_ORIENTATIONS = {
     126: "B",
     127: "B",
 }
-
-# Derived acquisition times that get stored in CDF data variables in L1a
-# processing. These are taken from the "Acq Time" column in the "Lo Stepping"
-# tab of the "*-SCI-LUT-*.xml" spreadsheet that largely defines CoDICE
-# processing.
-# TODO: Do away with this lookup table and instead calculate the acquisition
-#       times. See GitHub issue #1945.
-ACQUISITION_TIMES = {
-    0: [
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-    ],
-    1: [
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-    ],
-    2: [
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-    ],
-    3: [
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        578.70833333,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        289.35416667,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        192.90277778,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        144.67708333,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        115.74166667,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-        96.45138889,
-    ],
-}
-
-# These are for product that requires despinning in l1b.
-SW_INDEX_TO_POSITION = [1, 2, 3, 23, 24]
-NSW_INDEX_TO_POSITION = [
-    4,
-    5,
-    6,
-    7,
-    8,
-    9,
-    10,
-    11,
-    12,
-    13,
-    14,
-    15,
-    16,
-    17,
-    18,
-    19,
-    20,
-    21,
-    22,
-]
-
-# TODO: Update EFFICIENCY value when better information is available.
-# Constant for CoDICE Intensity calculations.
-EFFICIENCY = 1
 
 # Lookup table for mapping half-spin (keys) to esa steps (values)
 # This is used to determine geometry factors L2
