+++ imap_processing/tests/codice/test_codice_l2.py
@@ -12,6 +12,7 @@ from imap_data_access import AncillaryInput, ProcessingInputCollection
 from imap_processing import imap_module_directory
 from imap_processing.cdf.imap_cdf_manager import ImapCdfAttributes
 from imap_processing.cdf.utils import load_cdf, write_cdf
+from imap_processing.codice.codice_l1a import process_l1a
 from imap_processing.codice.codice_l1b import process_codice_l1b
 from imap_processing.codice.codice_l2 import (
     compute_geometric_factors,
@@ -21,7 +22,6 @@ from imap_processing.codice.codice_l2 import (
     process_lo_angular_intensity,
     process_lo_species_intensity,
 )
-from imap_processing.codice.codice_new_l1a import process_l1a
 from imap_processing.codice.constants import (
     LO_NSW_ANGULAR_VARIABLE_NAMES,
     LO_SW_ANGULAR_VARIABLE_NAMES,
